Explicitly track JS dependencies
This adds explicit tracking of JS dependencies used to transpile the TS
code in the repo for reproducibility. While at it, I also simplified the
way we generate the JS code by moving `package.json` from `scripts/`
into the root folder. This allows us to use npm to generate the code
instead of having introduce an additional dependency on `make`.